AUTH - message after creating a user
LESSONS
- getAllPractices improved as per roles
- deletePractice don't delete when lesson already finished
-deleteMyPractice created
-finishTime in model changed to string

+++ api/controllers/practice.controller.js
@@ -1,9 +1,15 @@
 const DriveLessonModel = require('../models/drivelesson.model')
 
 async function getAllPractices(req, res) {
+  console.log(res.locals.user.role)
   try {
-    const practices = await DriveLessonModel.find()
-    res.status(200).json(practices)
+    if (res.locals.user.role === 'teacher') {
+      const teacherPractices = await DriveLessonModel.find({ teacher: res.locals.user.id })
+      res.status(200).json(teacherPractices)
+    } else if (res.locals.user.role === 'admin') {
+      const practices = await DriveLessonModel.find()
+      res.status(200).json(practices)
+    }
   } catch (error) {
     res.status(500).send(`Request Error: ${error}`)
   }
@@ -20,8 +26,13 @@ async function updatePractice(req, res) {
 
 async function deletePractice(req, res) {
   try {
-    await DriveLessonModel.findByIdAndRemove(req.params.id)
-    res.status(200).send('Practice deleted successfully!')
+    const practice = await DriveLessonModel.findById(req.params.id)
+    if (practice.finishTime) {
+      res.status(200).send('Practice cannot be deleted!')
+    } else {
+      await DriveLessonModel.findByIdAndRemove(req.params.id)
+      res.status(200).send('Practice deleted successfully!')
+    }
   } catch (error) {
     res.status(500).send(`Request Error: ${error}`)
   }
